Begin developing national dtos
