+++ middleware/schema_and_dto/dtos/agencies/post.py
@@ -7,7 +7,7 @@ from middleware.schema_and_dto.dtos.agencies._helpers import (
     get_name_field,
     get_jurisdiction_type_field,
 )
-from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.core import (
+from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.generator.models.metadata import (
     MetadataInfo,
 )
