+++ middleware/schema_and_dto/dtos/data_requests/by_id/locations.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 
 from middleware.schema_and_dto.dtos.common.base import GetByIDBaseDTO
-from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.core import (
+from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.generator.models.metadata import (
     MetadataInfo,
 )
 from utilities.enums import SourceMappingEnum
