+++ middleware/schema_and_dto/dtos/signup.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field
 
-from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.core import (
+from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.generator.models.metadata import (
     MetadataInfo,
 )
