+++ middleware/schema_and_dto/dynamic_logic/pydantic_to_marshmallow/generator/models/field.py
@@ -1,0 +1,10 @@
+from marshmallow.fields import Field
+from pydantic import BaseModel, Field as PydanticField
+
+
+class MarshmallowFieldInfo(BaseModel):
+    class Config:
+        arbitrary_types_allowed = True
+
+    field: type[Field]
+    field_kwargs: dict = PydanticField(default_factory=dict)
