+++ tests/test_pydantic_to_marshmallow.py
@@ -5,6 +5,8 @@ from pydantic import BaseModel, Field
 
 from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.core import (
     generate_marshmallow_schema,
+)
+from middleware.schema_and_dto.dynamic_logic.pydantic_to_marshmallow.generator.models.metadata import (
     MetadataInfo,
 )
